:recycle: Rename fetch issues list
